# backend/openapi.yaml
openapi: 3.0.0
info:
  title: ReviseLikeTeacher API
  version: 1.0.0
  description: API for NEET PG preparation platform
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.reviseliketeacher.com
    description: Production server
paths:
  /auth/register:
    post:
      summary: Register new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '409':
          description: Email already exists
  /auth/login:
    post:
      summary: Login user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
  /auth/reset-password:
    post:
      summary: Request password reset
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Reset email sent
...
  /questions:
    get:
      summary: Get questions
      tags: [Questions]
      security:
        - bearerAuth: []
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: topic
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                type: object
                properties:
                  questions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Question'
                  total:
                    type: integer
  /questions/{id}:
    get:
      summary: Get question by ID
      tags: [Questions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Question details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
...
  /analytics/questions:
    get:
      summary: Per question analytics
      tags: [Analytics]
      security:
        - bearerAuth: []
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: subject
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Question analytics
...
  /admin/questions:
    get:
      summary: Get all questions (admin)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Questions list
    post:
      summary: Create question (admin)
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreate'
      responses:
        '201':
          description: Question created
  /admin/questions/{id}:
    put:
      summary: Update question (admin)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionUpdate'
      responses:
        '200':
          description: Question updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Question:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stem:
          type: string
        type:
          type: string
        subject:
          type: string
        topic:
          type: string
        difficulty:
          type: string
        importance:
          type: string
        cognitive_focus:
          type: string
        image_path:
          type: string
    QuestionCreate:
      type: object
      required: [stem, type, subject, topic]
      properties:
        stem:
          type: string
        type:
          type: string
        subject:
          type: string
        topic:
          type: string
        subtopic:
          type: string
        difficulty:
          type: string
        importance:
          type: string
        cognitive_focus:
          type: string
        ideal_answer:
          type: string
        key_points:
          type: array
          items:
            type: string
        previous_year_tags:
          type: array
          items:
            type: object
        image_path:
          type: string
    QuestionUpdate:
      type: object
      properties:
        stem:
          type: string
        type:
          type: string
        subject:
          type: string
        topic:
          type: string
        difficulty:
          type: string
        importance:
          type: string
        cognitive_focus:
          type: string
        status:
          type: string
...